// src/app/shared/ui/button/button.enums.ts
export enum ButtonSize {
    S = 'S',
    M = 'M',
    L = 'L',
    SUBSCRIBE = 'SUBSCRIBE',
    SEARCH = 'SEARCH',
}

export enum ButtonFillStyle {
    FILL = 'FILL',
    BORDER = 'BORDER',
    GHOST = 'GHOST',
    SUBSCRIBE = 'SUBSCRIBE',
}

export enum ButtonSpecialModifier {
    SEARCH = 'SEARCH',
    SUBSCRIBE = 'SUBSCRIBE',
}